Move SplitOffset to /shared/sliceutil
