+++ beacon-chain/core/helpers/shuffle.go
@@ -7,6 +7,7 @@ import (
 	"github.com/prysmaticlabs/prysm/shared/bytesutil"
 	"github.com/prysmaticlabs/prysm/shared/hashutil"
 	"github.com/prysmaticlabs/prysm/shared/params"
+	"github.com/prysmaticlabs/prysm/shared/sliceutil"
 )
 
 const seedSize = int8(32)
@@ -22,8 +23,8 @@ func SplitIndices(l []uint64, n uint64) [][]uint64 {
 	var divided [][]uint64
 	var lSize = uint64(len(l))
 	for i := uint64(0); i < n; i++ {
-		start := SplitOffset(lSize, n, i)
-		end := SplitOffset(lSize, n, i+1)
+		start := sliceutil.SplitOffset(lSize, n, i)
+		end := sliceutil.SplitOffset(lSize, n, i+1)
 		divided = append(divided, l[start:end])
 	}
 	return divided
@@ -123,19 +124,6 @@ func innerShuffledIndex(index uint64, indexCount uint64, seed [32]byte, shuffle 
 		}
 	}
 	return index, nil
-}
-
-// SplitOffset returns (listsize * index) / chunks
-//
-// Spec pseudocode definition:
-// def get_split_offset(list_size: int, chunks: int, index: int) -> int:
-//     """
-//     Returns a value such that for a list L, chunk count k and index i,
-//     split(L, k)[i] == L[get_split_offset(len(L), k, i): get_split_offset(len(L), k, i+1)]
-//     """
-//     return (list_size * index) // chunks
-func SplitOffset(listSize uint64, chunks uint64, index uint64) uint64 {
-	return (listSize * index) / chunks
 }
 
 // ShuffleList returns list of shuffled indexes in a pseudorandom permutation `p` of `0...list_size - 1` with ``seed`` as entropy.
