+++ shared/sliceutil/slice.go
@@ -272,3 +272,17 @@ func SplitCommaSeparated(arr []string) []string {
 	}
 	return result
 }
+
+// SplitOffset returns the start index of a given list splits into chunks,
+// it computes (listsize * index) / chunks.
+//
+// Spec pseudocode definition:
+// def get_split_offset(list_size: int, chunks: int, index: int) -> int:
+//     """
+//     Returns a value such that for a list L, chunk count k and index i,
+//     split(L, k)[i] == L[get_split_offset(len(L), k, i): get_split_offset(len(L), k, i+1)]
+//     """
+//     return (list_size * index) // chunks
+func SplitOffset(listSize uint64, chunks uint64, index uint64) uint64 {
+	return (listSize * index) / chunks
+}
